refactor(flatteners): split flatteners into separate files
